+++ CHANGELOG.md
@@ -28,6 +28,7 @@ QML Material CHANGELOG
  * [Popup] Support preventing a popup from being dismissed
  * [ActionBar] Control the overflow menu with Android hardware keys
  * [ActionButton] Add support for the mini and default FAB sizes
+ * [Dialog] Expose dialog buttons via aliases
 
 
 ### Version 0.0.6 (April 16, 2015)